move cache to pkg
