+++ pkg/cache/utils.go
